+++ src/rail/estimation/summarizer.py
@@ -4,8 +4,11 @@ Abstract base classes defining redshift estimations Informers and Estimators
 from rail.core.data import QPHandle, TableHandle, ModelHandle
 from rail.core.stage import RailStage
 
+from rail.estimation.informer import PzInformer
+# for backwards compatibility
 
-class CatSummarizer(RailStage):  #pragma: no cover
+
+class CatSummarizer(RailStage):  
     """The base class for classes that go from catalog-like tables
     to ensemble NZ estimates.
 
@@ -104,7 +107,7 @@ class PZSummarizer(RailStage):
         return self.get_handle('output')
 
 
-class SZPZSummarizer(RailStage):  #pragma: no cover
+class SZPZSummarizer(RailStage): 
     """The base class for classes that use two sets of data: a photometry sample with
     spec-z values, and a photometry sample with unknown redshifts, e.g. minisom_som and
     outputs a QP Ensemble with bootstrap realization of the N(z) distribution
@@ -187,57 +190,3 @@ class SZPZSummarizer(RailStage):  #pragma: no cover
         return self.get_handle('output')
 
 
-class PzInformer(RailStage):  #pragma: no cover
-    """The base class for informing models used to summarize photo-z posterior estimates
-    from ensembles of p(z) distributions.
-
-    PzSummarizers can use a generic "model", the details of which depends on the sub-class.
-    Some summaer will have associated PzInformer classes, which can be used to inform
-    those models.
-
-    (Note, "Inform" is more generic than "Train" as it also applies to algorithms that
-    are template-based rather than machine learning-based.)
-
-    PzInformer will produce as output a generic "model", the details of which depends on the sub-class.
-
-    They take as "input" a qp.Ensemble of per-galaxy p(z) data, which is used to "inform" the model.
-    """
-
-    name = 'Informer'
-    config_options = RailStage.config_options.copy()
-    inputs = [('input', QPHandle)]
-    outputs = [('model', ModelHandle)]
-
-    def __init__(self, args, comm=None):
-        """Initialize Informer that can inform models for redshift estimation """
-        RailStage.__init__(self, args, comm=comm)
-        self.model = None
-
-    def inform(self, training_data):
-        """The main interface method for Informers
-
-        This will attach the input_data to this `Informer`
-        (for introspection and provenance tracking).
-
-        Then it will call the run() and finalize() methods, which need to
-        be implemented by the sub-classes.
-
-        The run() method will need to register the model that it creates to this Estimator
-        by using `self.add_data('model', model)`.
-
-        Finally, this will return a ModelHandle providing access to the trained model.
-
-        Parameters
-        ----------
-        input_data :  `qp.Ensemble`
-            Per-galaxy p(z), and any ancilary data associated with it
-
-        Returns
-        -------
-        model : ModelHandle
-            Handle providing access to trained model
-        """
-        self.set_data('input', training_data)
-        self.run()
-        self.finalize()
-        return self.get_handle('model')
